Rename ChromeTabButton to ConnectLoginButton
