+++ connect-id-example/src/main/java/com/telenor/connect/connectidexample/SignInActivity.java
@@ -8,13 +8,13 @@ import android.view.View;
 
 import com.telenor.connect.ConnectCallback;
 import com.telenor.connect.ConnectSdk;
-import com.telenor.connect.ui.ConnectLoginChromeTabButton;
+import com.telenor.connect.ui.ConnectLoginButton;
 import com.telenor.connect.utils.ConnectUtils;
 
 public class SignInActivity extends Activity {
 
     private View progressBar;
-    private ConnectLoginChromeTabButton loginButton;
+    private ConnectLoginButton loginButton;
 
     @Override
     protected void onResume() {
@@ -33,7 +33,7 @@ public class SignInActivity extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_sign_in);
-        loginButton = (ConnectLoginChromeTabButton) findViewById(R.id.login_button);
+        loginButton = (ConnectLoginButton) findViewById(R.id.login_button);
         progressBar = findViewById(R.id.progress_bar);
 
         loginButton.setLoginScopeTokens("profile openid");
@@ -50,12 +50,7 @@ public class SignInActivity extends Activity {
         ConnectSdk.checkIntentForAndHandleRedirectUrlCall(getIntent(), new ConnectCallback() {
             @Override
             public void onSuccess(Object successData) {
-                final Intent startIntent
-                        = new Intent(getApplicationContext(), SignedInActivity.class);
-                startIntent.setFlags(
-                        Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
-                startActivity(startIntent);
-                finish();
+                goToSignedInActivity();
             }
 
             @Override
@@ -65,13 +60,20 @@ public class SignInActivity extends Activity {
         });
     }
 
+    private void goToSignedInActivity() {
+        final Intent intent
+                = new Intent(getApplicationContext(), SignedInActivity.class);
+        intent.setFlags(
+                Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        startActivity(intent);
+        finish();
+    }
+
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         if (resultCode == Activity.RESULT_OK) {
-            Intent intent = new Intent(getApplicationContext(), SignedInActivity.class);
-            startActivity(intent);
-            finish();
+            goToSignedInActivity();
         }
     }
 }
