+++ connect/src/com/telenor/connect/ui/ConnectLoginButton.java
@@ -1,6 +1,15 @@
 package com.telenor.connect.ui;
 
+import android.content.ComponentName;
 import android.content.Context;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
+import android.support.annotation.NonNull;
+import android.support.customtabs.CustomTabsClient;
+import android.support.customtabs.CustomTabsIntent;
+import android.support.customtabs.CustomTabsServiceConnection;
+import android.support.customtabs.CustomTabsSession;
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.view.View;
@@ -8,88 +17,109 @@ import android.view.View;
 import com.telenor.connect.ConnectException;
 import com.telenor.connect.ConnectSdk;
 import com.telenor.connect.R;
-import com.telenor.connect.id.Claims;
 import com.telenor.connect.utils.ClaimsParameterFormatter;
 import com.telenor.connect.utils.Validator;
 
 import org.json.JSONException;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
-public class ConnectLoginButton extends ConnectButton {
+public class ConnectLoginButton extends ConnectWebViewLoginButton {
 
-    public static final int NO_CUSTOM_LAYOUT = -1;
-
-    private ArrayList<String> acrValues;
-    private Map<String, String> loginParameters;
-    private ArrayList<String> loginScopeTokens;
-    private int requestCode = 0xa987;
-    private Claims claims;
-    private int customLoadingLayout = NO_CUSTOM_LAYOUT;
+    private OnClickListener onClickListener;
+    private Uri uri;
+    private CustomTabsClient customTabsClient;
+    private CustomTabsSession customTabsSession;
+    private boolean shouldPreLoad = false;
+    private boolean serviceConnected = false;
+    private boolean customTabsSupported = false;
 
     public ConnectLoginButton(Context context, AttributeSet attributeSet) {
         super(context, attributeSet);
         setText(R.string.com_telenor_connect_login_button_text);
-        setOnClickListener(new LoginClickListener());
+
+        CustomTabsServiceConnection connection = new CustomTabsServiceConnection() {
+            @Override
+            public void onCustomTabsServiceConnected(ComponentName name, CustomTabsClient client) {
+                customTabsClient = client;
+                customTabsClient.warmup(0);
+                customTabsSession = customTabsClient.newSession(null);
+                if (shouldPreLoad && customTabsSession != null) {
+                    customTabsSession.mayLaunchUrl(uri, null, null);
+                }
+                serviceConnected = true;
+            }
+
+            @Override
+            public void onServiceDisconnected(ComponentName name) {
+                serviceConnected = false;
+            }
+        };
+
+        if (!serviceConnected) {
+            customTabsSupported = CustomTabsClient.bindCustomTabsService(
+                    getContext(), "com.android.chrome", connection);
+        }
+        onClickListener = new LoginClickListener();
+        setOnClickListener(onClickListener);
     }
 
-    public ArrayList<String> getAcrValues() {
-        return acrValues;
+    public OnClickListener getOnClickListener() {
+        return onClickListener;
     }
 
-    public Map<String, String> getLoginParameters() {
-        return loginParameters;
+    public void preLoad() {
+        if (uri == null) {
+            uri = getAuthorizeUriAndSetLastAuthState();
+        }
+        if (customTabsSession == null) {
+            shouldPreLoad = true;
+        } else {
+            customTabsSession.mayLaunchUrl(uri, null, null);
+        }
     }
 
-    public ArrayList<String> getLoginScopeTokens() {
-        return loginScopeTokens;
+    private Uri getAuthorizeUriAndSetLastAuthState() {
+        final Map<String, String> parameters = getParameters();
+        return ConnectSdk.getAuthorizeUriAndSetLastAuthState(parameters);
     }
 
-    public int getRequestCode() {
-        return requestCode;
+    @NonNull
+    private Map<String, String> getParameters() {
+        Map<String, String> parameters = new HashMap<>();
+        if (getAcrValues() != null && !getAcrValues().isEmpty()) {
+            parameters.put("acr_values", TextUtils.join(" ", getAcrValues()));
+        }
+
+        if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
+            parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
+        }
+
+        if (getClaims() != null && getClaims().getClaimsAsSet() != null) {
+            try {
+                parameters.put("claims", ClaimsParameterFormatter.asJson(getClaims()));
+            } catch (JSONException e) {
+                throw new ConnectException(
+                        "Failed to create claims Json. claims="+getClaims(), e);
+            }
+        }
+
+        if (getLoginParameters() != null && !getLoginParameters().isEmpty()){
+            parameters.putAll(getLoginParameters());
+        }
+        return parameters;
     }
 
-    public Claims getClaims() {
-        return claims;
-    }
-
-    public int getCustomLoadingLayout() {
-        return customLoadingLayout;
-    }
-
-    public void setAcrValues(String... acrValues) {
-        this.acrValues = new ArrayList<>(Arrays.asList(acrValues));
-    }
-
-    public void setAcrValues(ArrayList<String> acrValues) {
-        this.acrValues = acrValues;
-    }
-
-    public void setLoginScopeTokens(String... scopeTokens) {
-        loginScopeTokens = new ArrayList<>(Arrays.asList(scopeTokens));
-    }
-
-    public void setLoginScopeTokens(ArrayList<String> scopeTokens) {
-        loginScopeTokens = scopeTokens;
-    }
-
-    public void addLoginParameters(Map<String, String> parameters) {
-        loginParameters = parameters;
-    }
-
-    public void setRequestCode(int requestCode) {
-        this.requestCode = requestCode;
-    }
-
-    public void setClaims(Claims claims) {
-        this.claims = claims;
-    }
-
-    public void setCustomLoadingLayout(int customLoadingLayout) {
-        this.customLoadingLayout = customLoadingLayout;
+    private boolean contextIntentFilterMatchesRedirectUri() {
+        final Uri parse = Uri.parse(ConnectSdk.getRedirectUri());
+        final Intent intent = new Intent().setData(parse);
+        final PackageManager manager = getContext().getPackageManager();
+        final ComponentName componentName = intent.resolveActivity(manager);
+        if (componentName == null) {
+            return false;
+        }
+        return getContext().getClass().getName().equals(componentName.getClassName());
     }
 
     private class LoginClickListener implements OnClickListener {
@@ -97,36 +127,25 @@ public class ConnectLoginButton extends ConnectButton {
         @Override
         public void onClick(View v) {
             Validator.sdkInitialized();
-
-            Map<String, String> parameters = new HashMap<>();
-            if (getAcrValues() != null && !getAcrValues().isEmpty()) {
-                parameters.put("acr_values", TextUtils.join(" ", getAcrValues()));
-            }
-
-            if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
-                parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
-            }
-
-            if (claims != null && claims.getClaimsAsSet() != null) {
-                try {
-                    parameters.put("claims", ClaimsParameterFormatter.asJson(claims));
-                } catch (JSONException e) {
-                    throw new ConnectException("Failed to create claims Json. claims="+claims, e);
+            if (!customTabsSupported || !contextIntentFilterMatchesRedirectUri()) {
+                int customLoadingLayout = getCustomLoadingLayout();
+                if (customLoadingLayout == NO_CUSTOM_LAYOUT) {
+                    ConnectSdk.authenticate(getActivity(), getParameters(), getRequestCode());
+                } else {
+                    ConnectSdk.authenticate(
+                            getActivity(), getParameters(), customLoadingLayout, getRequestCode());
                 }
+
+                return;
             }
 
-            if (getLoginParameters() != null && !getLoginParameters().isEmpty()){
-                parameters.putAll(getLoginParameters());
+            if (uri == null) {
+                uri = getAuthorizeUriAndSetLastAuthState();
             }
 
-            if (customLoadingLayout == NO_CUSTOM_LAYOUT) {
-                ConnectSdk.authenticate(getActivity(), parameters, requestCode);
-            } else {
-                ConnectSdk.authenticate(getActivity(),
-                        parameters,
-                        customLoadingLayout,
-                        requestCode);
-            }
+            new CustomTabsIntent.Builder()
+                    .build()
+                    .launchUrl(getActivity(), uri);
         }
     }
 }
