+++ connect/tests/src/main/res/layout/com_telenor_test_activity_layout.xml
@@ -2,7 +2,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
              android:layout_width="match_parent"
              android:layout_height="match_parent">
-    <com.telenor.connect.ui.ConnectLoginButton
+    <com.telenor.connect.ui.ConnectWebViewLoginButton
         android:id="@+id/login_button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
