+++ connect/tests/src/test/java/com/telenor/connect/ui/ConnectWebViewLoginButtonTest.java
@@ -24,7 +24,7 @@ import static org.robolectric.Shadows.shadowOf;
 
 @RunWith(RobolectricTestRunner.class)
 @Config(manifest = "src/main/AndroidManifest.xml", sdk = 18)
-public class ConnectLoginButtonTest {
+public class ConnectWebViewLoginButtonTest {
 
     @Before
     public void before() {
@@ -34,7 +34,7 @@ public class ConnectLoginButtonTest {
     @Test(expected = ConnectNotInitializedException.class)
     public void clickingLoginButtonBeforeInitializingSdkThrows() {
         Activity activity = Robolectric.buildActivity(TestActivity.class).create().get();
-        ConnectLoginButton button = (ConnectLoginButton) activity.findViewById(R.id.login_button);
+        ConnectWebViewLoginButton button = (ConnectWebViewLoginButton) activity.findViewById(R.id.login_button);
         button.performClick();
     }
 
@@ -42,7 +42,7 @@ public class ConnectLoginButtonTest {
     public void clickingLoginButtonWithInitializedSdkStartsConnectActivity() {
         ConnectSdk.sdkInitialize(RuntimeEnvironment.application);
         Activity activity = Robolectric.buildActivity(TestActivity.class).create().get();
-        ConnectLoginButton button = (ConnectLoginButton) activity.findViewById(R.id.login_button);
+        ConnectWebViewLoginButton button = (ConnectWebViewLoginButton) activity.findViewById(R.id.login_button);
         button.setLoginScopeTokens("profile");
         button.performClick();
